requirements.txt
dspy-ai
langtrace-python-sdk
loguru